Avatar For Source and User | add data (type - Key - Secret ) to Table sources | Create Module Zones |  Add Column Color to table Order_status

+++ api/Modules/User/Http/Controllers/Roles/RoleController.php
@@ -6,16 +6,25 @@ use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Modules\User\Entities\User;
-use Illuminate\Routing\Controller;
 use Spatie\Permission\Models\Role;
 use Spatie\QueryBuilder\QueryBuilder;
 use Spatie\QueryBuilder\AllowedFilter;
 use App\Http\Controllers\ApiController;
 use Modules\User\Http\Requests\StoreRoleRequest;
 use Modules\User\Http\Requests\UpdateRoleRequest;
+use Modules\User\Transformers\Roles\RoleCollection;
+use Modules\User\Transformers\Roles\RoleResource;
 
 class RoleController extends ApiController
 {
+
+    private $user;
+    public function __construct()
+    {
+        $this->user = auth('sanctum')->user();
+        $this->middleware('role:super-admin|admin');
+    }
+
     // Index
     public function index(Request $request)
     {
@@ -24,23 +33,40 @@ class RoleController extends ApiController
         $offset = $request->query->get('offset') ?? 0;
         //
         $columns = [
-            AllowedFilter::exact('id'), 'name', 'created_at',
+            AllowedFilter::exact('id'), 'name', 'created_at', 'updated_at',
         ];
+
         //
-        $roles = QueryBuilder::for(Role::class)
-            //filtering
-            ->allowedFilters($columns)
-            //sorting
-            ->defaultSort('id')
-            ->allowedSorts([
-                // Products
-                'id', 'name', 'created_at',
-            ])
-            ->limit($limit)->offset($offset * $limit)
-            ->get() ?? [];
+        if ($this->user->hasRole('super-admin')) { // Super-admin
+            $roles = QueryBuilder::for(Role::class)
+                //filtering
+                ->allowedFilters($columns)
+                //sorting
+                ->defaultSort('id')
+                ->allowedSorts([
+                    // Products
+                    'id', 'name', 'created_at', 'updated_at',
+                ])
+                ->limit($limit)->offset($offset * $limit)
+                ->get() ?? [];
+        } elseif ($this->user->hasRole('admin')) {  // Admin
+            $roles = QueryBuilder::for(Role::class)
+                //filtering
+                ->allowedFilters($columns)
+                //sorting
+                ->defaultSort('id')
+                ->allowedSorts([
+                    // Products
+                    'id', 'name', 'created_at', 'updated_at',
+                ])
+                ->limit($limit)->offset($offset * $limit)
+                ->whereIn('name', ['manager', 'member'])->get() ?? [];
+        }
+
         //
-        return $this->showAll($roles, Response::HTTP_OK);
+        return $this->showAll(new RoleCollection($roles), Response::HTTP_OK);
     }
+
     // Show
     public function show(Role $role)
     {
@@ -52,12 +78,19 @@ class RoleController extends ApiController
         //
         return $this->errorResponse(__('messages.recource_cannot_be_found'), Response::HTTP_NOT_FOUND);
     }
+
     // Store
     public function store(StoreRoleRequest $request)
     {
-        $role = Role::create(['name' => $request->get('role_name')]);
-        return $this->successResponse($role, Response::HTTP_CREATED);
+
+        if ($this->user->hasRole('super-admin')) {
+            $role = Role::create(['name' => $request->get('role_name')]);
+            return $this->successResponse(new RoleResource($role), Response::HTTP_CREATED);
+        } else {
+            return $this->errorResponse('UNAUTHORIZED', Response::HTTP_UNAUTHORIZED);
+        }
     }
+
     // Get Role by user
     public function getRolesByUser(User $user)
     {
@@ -74,24 +107,25 @@ class RoleController extends ApiController
             return $this->errorResponse($ex->getMessage(), Response::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
+
+
     // User has role
-    public function userHasRole(User $user, $roleName)
+    public function userHasRole(User $user, Role $role)
     {
         try {
-            if (!empty($user)) {
-                $hasRole = $user->hasRole($roleName);
-                return $this->successResponse($hasRole, Response::HTTP_OK);
-            }
-            return $this->errorResponse(trans('messages.resource_cannot_be_found_in_database'), Response::HTTP_NOT_FOUND);
+            $hasRole = $user->hasRole($role->name);
+            return $this->successResponse(['has_role ' => $hasRole], Response::HTTP_OK);
         } catch (Exception $ex) {
             return $this->errorResponse($ex->getMessage(), Response::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
+
+
     // Update
     public function update(UpdateRoleRequest $request, Role $role)
     {
         //
-        if ($role) {
+        if ($this->user->hasRole('super-admin')) {
             // name
             if (request()->has('role_name')) {
                 $role->name = $request->role_name;
@@ -99,24 +133,30 @@ class RoleController extends ApiController
             $role->save();
             //
             return $this->successResponse(__('messages.updated_successfully'), Response::HTTP_NO_CONTENT);
+        } else {
+            return $this->errorResponse('UNAUTHORIZED', Response::HTTP_UNAUTHORIZED);
         }
-        //
-        return $this->errorResponse(__('messages.recource_cannot_be_found'), Response::HTTP_NOT_FOUND);
     }
+
+
     // Destroy
     public function destroy(Role $role)
     {
         //
-        if ($role) {
-            //
-            if ($role->delete()) {
+        if ($this->user->hasRole('super-admin')) {
+            if ($role) {
                 //
-                return $this->successResponse(__('messages.deleted_successfully'), Response::HTTP_NO_CONTENT);
+                if ($role->delete()) {
+                    //
+                    return $this->successResponse(__('messages.deleted_successfully'), Response::HTTP_NO_CONTENT);
+                }
+                //
+                return $this->errorResponse(__('messages.resource_cannot_be_deleted'), Response::HTTP_CONFLICT);
             }
             //
-            return $this->errorResponse(__('messages.resource_cannot_be_deleted'), Response::HTTP_CONFLICT);
+            return $this->errorResponse(__('messages.recource_cannot_be_found'), Response::HTTP_NOT_FOUND);
+        } else {
+            return $this->errorResponse('UNAUTHORIZED', Response::HTTP_UNAUTHORIZED);
         }
-        //
-        return $this->errorResponse(__('messages.recource_cannot_be_found'), Response::HTTP_NOT_FOUND);
     }
 }
